+++ lily.h
@@ -6,7 +6,7 @@
 
 void write_lilyhead(FILE *lilyfile, char *name);
 
-void print_note(MappingArray *durs, FILE *lilyfile, char *note, int duration);
+void print_note(MappingArray *durs, FILE *lilyfile, char *note, double duration);
 
 void write_lilytail(FILE *lilyfile, int metronome);
